Fixup multiple passes
Don't barf on a second pass this is mostly useful for testing changes.
For a full build you should clean the whole work directory before
running.

+++ piclock-customisations/03-symlink-config/00-run-chroot.sh
@@ -8,6 +8,8 @@ if [ ! -d "$SD_CFG_DIR" ]; then
 fi
 
 for a in ntpsec/ntp.conf piclock.cfg network/interfaces.d/wired-network.conf; do
-	mv "/etc/$a" "$SD_CFG_DIR"
-	ln -s "$SD_CFG_DIR/$(basename "$a")" "/etc/$a"
+	if [ ! -L "/etc/$a" ]; then
+		mv "/etc/$a" "$SD_CFG_DIR"
+		ln -s "$SD_CFG_DIR/$(basename "$a")" "/etc/$a"
+	fi
 done
